+++ src/Data/MotoBest.Scraping/Common/AdvertScrapeModel.cs
@@ -28,7 +28,7 @@
 
         public int HorsePowers { get; set; }
 
-        public int Kilometrage { get; set; }
+        public long Kilometrage { get; set; }
 
         public int Views { get; set; }
